chore: move 'createAndFundNewWalletForPhone' to helpers module
